Change error category message to also work with scoped enums

+++ include/result/basic_error_category.inl
@@ -18,7 +18,7 @@ namespace ol {
     template<const char* Name, typename ErrEnum, std::string (*DescFn)(int), int (*ErrcMapFn)(int)>
 	std::string 
     basic_error_category<Name, ErrEnum, DescFn, ErrcMapFn>::message(int c) const {
-		if (c < 0 || c >= ErrEnum::num_codes) 
+		if (c < 0 || c >= static_cast<int>(ErrEnum::num_codes)) 
 			std::string("<unknown error code; no message available>");
 
 		std::string msg = impl::category_fn<DescFn, ErrcMapFn>::get_desc(c);
